Live Data functionality added

+++ time_convert.py
@@ -1,6 +1,3 @@
-import numpy as np
-
-
 def double_digit_convert(str):
     if len(str) == 1:
         double_digit_string = "0" + str
@@ -28,4 +25,4 @@ def time_strip(start_date, start_time, end_date, end_time):
 
     start = start_year + start_month + start_day + start_hour + start_minute + start_second
     end = end_year + end_month + end_day + end_hour + end_minute + end_second
-    return (np.int64(start), np.int64(end))
+    return (int(start), int(end))
